--the CategorizedLinePool class now has a function to solved for unsolved interceptsPointPrecise values (CategorizedLinePool::solveUnsolvedInterceptsPointPrecise)
--did several conversions of map size to int, to make compiler happy and remove excessive warnings.

--added new function, CleaveSequenceFactory::printLineCounts()

--more work on CoplanarCategorizedLineComparison::performLineComparison()

--the function CoplanarMassCreator::runMassManipulation() will now call clearLinePools() for the trackedCopy's sequence factory.

--added new class, UnsolvedPreciseSolver. This class will find the border lines that would be produced for a CategorizedLine type of INTERCEPTS_POINT_PRECISE; it is used as part of the process to convert lines of the IntersectionType::INTERCEPTS_POINT_PRECISE_UNSOLVED_COPLANAR to this value.

--the single existing call of the function, CleaveSequenceFactory::clipTwinCategorizedLinesofInterceptPointPrecise() has been removed/commented out for the time being; this function needs to be reworked.

+++ OrganicGLWinLib/LineSegmentIntersectAnalyzer.cpp
@@ -246,7 +246,7 @@ bool TwoDLineSegmentIntersectAnalyzer::checkForColinearOverlap()
 TwoDPoint TwoDLineSegmentIntersectAnalyzer::round2DPointToHundredths(TwoDPoint in_point)
 {
 	TwoDPoint returnPoint = in_point;
-	returnPoint.x = floor(in_point.x * 100 + 0.5) / 100;
-	returnPoint.y = floor(in_point.y * 100 + 0.5) / 100;
+	returnPoint.x = float(floor(in_point.x * 100 + 0.5) / 100);
+	returnPoint.y = float(floor(in_point.y * 100 + 0.5) / 100);
 	return returnPoint;
 }
